test/unit: use posix_memalign() to allocate reactor
As mentioned in 771a7b879, there seems to be a bug with gcc and asan
where it doesn't properly align structures declared on the stack.

It also looks like declaring the variable with the aligned attribute
doesn't work either, so just use posix_memalign() to work around this.

Fixes #3004.

Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: I4af902af0384cbd429334433c291ead23eba8293
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/18855
Community-CI: Mellanox Build Bot
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Reviewed-by: Jim Harris <james.r.harris@intel.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Shuhei Matsumoto <smatsumoto@nvidia.com>
Reviewed-by: Kamil Godzwon <kamilx.godzwon@intel.com>