examples/perf: Avoid losing references to perf tasks.
Requeuing perf tasks for submission is done by swapping
currently queued tasks to a local list and then submitting
them one by one with TAILQ_FOREACH() and submit_single_io().
The problem here is that upon further failure in
submit_single_io() we move a task back to the queued list,
which causes the original list, on which TAILQ_FOREACH()
operates, to lose reference to the second element and exit
the loop. So in the end we just move the first task and the
rest are left hanging in the memory.

Fixes #3032

Change-Id: Ia55ff448be4213b9bfe3a685ec14b097f26a2278
Signed-off-by: Krzysztof Karas <krzysztof.karas@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/18908
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>