accel: add seq pointer to task data type
Accel driver might need to internally queue the sequences. However,
sequence type is opaque and there is no access to its internal link.
By adding sequence pointer to task data type (which is already public)
task's link can be used to achive that.

Exposing sequence type to public is problematic due to few factors.
Mostly it is because it uses few data types which are not yet public
and due to general rule that implemenation details shall be hidden
from the end accel users to not create unnecessary dependencies.

Signed-off-by: Jacek Kalwas <jacek.kalwas@intel.com>
Change-Id: I4db8009929909ffb58dff77d0a7c59e1e9ac82a9
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/18922
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Aleksey Marchuk <alexeymar@nvidia.com>
Community-CI: Mellanox Build Bot