test/common: change asan/ubsan error exit codes
By default asan/ubsan set exit status to 1 if an error is encountered,
which is fine in most cases.  However, when an app is expected to fail,
it's usually started with `NOT foo`, which will succeed if the program
exits with 1, meaning that it won't catch asan/ubsan errors.  So, change
the exit status to a value that translates to SIGABRT, causing `NOT foo`
to fail.

Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: I6e51aef0de799a3adccd5a2853a5be2f96bacd4a
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/18992
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Jim Harris <jim.harris@gmail.com>
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Reviewed-by: Sebastian Brzezinka <sebastian.brzezinka@intel.com>
Reviewed-by: Michal Berger <michal.berger@intel.com>