go/rpc: Removal of spdk.go file
This is the original commit message with this file creation:
"The only purpose at the moment is to allow vendoring of the SPDK
source code into Go projects which manage dependencies with the "dep"
tool. That tool is designed for Go projects and as a sanity check
rejects any dependency which does not have at least one Go file, at
least at the moment"

Since we introduced new Go files there is no need to keep this file anymore.

Change-Id: Ia6f56a8e56a14f6ef00a2302002fb21f0832a95e
Signed-off-by: Maciej Mis <maciej.mis@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/19012
Reviewed-by: Sebastian Brzezinka <sebastian.brzezinka@intel.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Ben Walker <ben@nvidia.com>
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Community-CI: Mellanox Build Bot