module/raid: rebuild support
Add support for running a background "process" for performing the
rebuild I/O. In the future, different types of processes can be added,
like resync or migration.

The process performs its work inside a "window" - a range of LBAs of
predefined maximum size. The basic flow of the process is:
- lock (quiesce) the range
- submit process requests, handled by the raid module implementation
- completion of process requests
- update io channels - move the processed range offset
- unlock (unquiesce) the range
- repeat until all blocks are processed or the process is stopped

Change-Id: I6a89bf06791683cac2e178211efe7b66a34394eb
Signed-off-by: Artur Paszkiewicz <artur.paszkiewicz@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/18739
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Jim Harris <jim.harris@samsung.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Reviewed-by: LiYankun <845245370@qq.com>