lib/accel: Reapck & reorder accel_task
1. Change type of op_code tp uint8
2. Change type of status to int16
These chanes a bit reduces size of a hole and
size of structure:
Origianl accel task:
        /* size: 384, cachelines: 6, members: 22 */
        /* sum members: 380, holes: 1, sum holes: 4 */
With this change:
        /* size: 376, cachelines: 6, members: 22 */
        /* sum members: 375, holes: 1, sum holes: 1 */
        /* last cacheline: 56 bytes */

Main advanatage of this change is that "heavy" members
such as bounce and aux_iovs are now at the end of the
structure. These member are not used with each task,
e.g. bounce is only used when accel module doesn't
support memory domain.
All member which are almost always used in IO path
occupy first 4 cache lines

Signed-off-by: Alexey Marchuk <alexeymar@nvidia.com>
Change-Id: Ib5215723e838f96edb983bf7dff71fe82eb4c8a1
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/19037
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Community-CI: Mellanox Build Bot