lib/blob: Clear cluster before releasing due to unneeded allocation
When we allocate a cluster, we may copy parent data to the new cluster.
After this copy is done, we may realize that another IO channel already
allocated the cluster so we no longer need the cluster. We need to
release it. But we need to clear the data. Otherwise, another blob may
claim the cluster and may see garbage data instead of zeroes.

Change-Id: Ib3fb9d4a264371bcaa224855ad0c3c1bbdc93979
Signed-off-by: Alex Michon <amichon@kalrayinc.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/19231
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Jim Harris <jim.harris@samsung.com>
Community-CI: Mellanox Build Bot