nvme/tcp: force qpair poll only when there are queued reqs
Completing an accel crc32 data digest calculation might cause a request
to be completed outside of the context of process_completions().  In
this case, a qpair needs to be put on the needs_poll list to make sure
that any queued requests are resubmitted.  There's no point in doing so
if there are no queued requests, so the condition was changed to reflect
that.

Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: I0000fe509fe62179180f6b0b4e56ae78e2214c05
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/19265
Reviewed-by: Aleksey Marchuk <alexeymar@nvidia.com>
Community-CI: Mellanox Build Bot
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Ben Walker <benwalker@nvidia.com>