test/check_so_deps: refactor pulling spdk-abi repo
This patch refactors pulling spdk-abi repository for
confirm_abi_deps test.

Main goal is to make it easier to developers to execute
check_so_deps.sh, by allowing to have a pre-existing
spdk-abi repository on their system and point to it.
Rather than hide it in SPDK_ABI_DIR variable passed in
autotest config file.

Now if second argument is provided it will be used as
path to spdk-abi. This repository will not be removed
at the end of the test.

If not provided, the repo will be cloned and then deleted
at the end of the test.

CI used to pass the path via SPDK_ABI_DIR, so it is now
included as second parameter when called from autobuild.

Change-Id: I8110c2bf4b25658d7f7f6758d00ea57e3ea6e60a
Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/19275
Reviewed-by: Ben Walker <benjamin.walker@intel.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>