bdev/nvme: wait for detached nvme controller in rpc
rpc_bdev_nvme_detach_controller did not wait for an operation
completion. If a sequence attach-detach-attach with the same
name is run, the script might fail if the detach is not
complete on an active path before the second attach.
Wait for a path does not exist before sending back a json rpc
response.

Change-Id: Id0e1fb49e69745acf1479188307aba20bf1b2020
Signed-off-by: Artsiom Koltun <artsiom.koltun@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/19340
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Reviewed-by: Shuhei Matsumoto <smatsumoto@nvidia.com>
Community-CI: Mellanox Build Bot
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>