test/unit: convert blob_ut to use spdk_ut_run_tests()
Converting this test was a little bit more involved, as it runs all
tests multiple times with different parameter combinations.  Now, a
separate test suite is registered for each such combination, which makes
it compatible with spdk_ut_run_tests() and allows users to run a given
test case (or test suite) with specific parameters.

Also, this patch fixes a bug when some of the test failures weren't
reported in the exit status of the process.

Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: I1a9fcbc370cc2cdfe366255f56da6df950e2aa59
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/19320
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Shuhei Matsumoto <smatsumoto@nvidia.com>
Reviewed-by: Aleksey Marchuk <alexeymar@nvidia.com>
Community-CI: Mellanox Build Bot