lib/env_dpdk: update PCI headers for DPDK 23.07
The header copies are placed in directory named
after first release where a breaking change was introduced.
DPDK 22.11 introduced a breaking change and as such new
copies were created.
Since then, DPDK 23.07 is the first version where API is changed.
This time around it does not affect backwards compatibility,
as new functions were added. Headers from DPDK 22.11 will still
work for DPDK 23.07 and prior versions, so long as the new functions
are not used.

To avoid unnecessarily creating another copy of headers and implementation,
for DPDK 23.07, the DPDK 22.11 remained as they were.

If rte_pci_mmio_read/write() are to be used in future,
simply a new copy and implementation should be created.
To catch the case of someone unaware of this versioning,
starts to use this API an redefinition of those functions were
added.

For purposes of test verifying the DPDK PCI headers,
a patch removing those functions was added.
This way DPDK 23.07 and later will match current
header copies.

Change-Id: I5d3813eced7f15d6dc4d7b8654f256ae1eebcd32
Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/19333
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Sebastian Brzezinka <sebastian.brzezinka@intel.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Reviewed-by: Changpeng Liu <changpeng.liu@intel.com>