test/env_dpdk: add starting version for DPDK patches
There are cases that we might want to apply different
patches to DPDK the same copy of the headers. This
depends on which DPDK is currently tested against.

To support that, the patches should be prepended
with first DPDK version that it applies.

Next patch in series will add new patch for DPDK 23.07,
yet the header copies will remain the same - DPDK 22.11.
Then particular patch will be applied when testing
DPDK 23.07 or later.

Change-Id: I365fabfd161fd09f72fc5e8fbe1ccefa99648e4c
Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/19360
Reviewed-by: Sebastian Brzezinka <sebastian.brzezinka@intel.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Changpeng Liu <changpeng.liu@intel.com>