test/env_dpdk: add starting version for DPDK patches
There are cases that we might want to apply different
patches to DPDK the same copy of the headers. This
depends on which DPDK is currently tested against.

To support that, the patches should be prepended
with first DPDK version that it applies.

Next patch in series will add new patch for DPDK 23.07,
yet the header copies will remain the same - DPDK 22.11.
Then particular patch will be applied when testing
DPDK 23.07 or later.

Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/19360 (master)

(cherry picked from commit 11694e8d59b8b109194a6acebefd0506ba92d338)
Change-Id: I365fabfd161fd09f72fc5e8fbe1ccefa99648e4c
Signed-off-by: Krzysztof Karas <krzysztof.karas@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/19827
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>