nvme: respect abort_dnr when aborting queued req in submit
When a qpair is disconnected, queued requests are aborted with DNR bit
set to abort_dnr.  This wasn't true if we detected that a qpair was
disconnected while submitting a queued request.  Additionally, such
requests will now be completed with SPDK_NVME_SC_ABORTED_SQ_DELETION to
keep the behavior aligned with those aborted when qpair is disconnected.

Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: I2a4df203a9afe7fc19ca3e7258f7aa88ac5812de
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/19445
Community-CI: Mellanox Build Bot
Reviewed-by: Aleksey Marchuk <alexeymar@nvidia.com>
Reviewed-by: Jim Harris <jim.harris@gmail.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>