test/vhost: Don't use RDMA for migration-TC2
This idea behind this test case is not to focus on a particular
transport protocol but rather use of nvmf during migration in
general. Since RDMA requires specific hardware, it imposes
unnecessary limit on the CI pool as to where this test suite can
be executed. Avoid that and use TCP for which fallback to veth
setup is supported in case actual hw is not present.

Change-Id: I6ec3f701285e26e42f5d5cbdbc078a5d57778797
Signed-off-by: Michal Berger <michal.berger@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/19541
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Reviewed-by: Jim Harris <jim.harris@gmail.com>