test/vhost: Perform verify only for the write job
Since fio-3.33, due to the following commit (bisect + revert):

  https://github.com/axboe/fio/commit/191d6634e8a692bef15143715c88920987ecaa89

using verification for *read jobs is foobar. For currently existing
configuration, fio >= 3.33 exits almost immediately causing the
migration test to fail - before VM is migrated, we check if fio is
actually running, however, due to this issue, it's not.

To mitigate, verify only the write job, allowing the randread job
to block for expected amount of time (20s).

Change-Id: I7c36e8f56635292426a15aeeb2395cecd454058a
Signed-off-by: Michal Berger <michal.berger@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/19543
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Jim Harris <jim.harris@gmail.com>