nvme/tcp: check PDU's data len before executing a sequence
If a CapsuleCmd is sent without in-capsule data, we should wait before
executing accel sequence until sending the first H2CData PDU to allow the
data digest calculation to be executes as part of the existing sequence.
Otherwise, it would be calculated separately, which can be inefficient
if it was possible to chain it with previous operations in the sequence.

Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: Ic5b791736bef1f8a937be918e99c6babd8ce00aa
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/19547
Reviewed-by: Changpeng Liu <changpeng.liu@intel.com>
Reviewed-by: Jim Harris <jim.harris@gmail.com>
Community-CI: Mellanox Build Bot
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>