test/iscsi_tgt: Simplify waitfortcp()
Remove extra util checks and stick to procfs. In practice, this test
is bound to linux under our CI so there's no point in having an
explicit checks for it (e.g. the ip is not available under freebsd
and its netstat's output differs from what scripts expects to find,
i.e., it looks for linux's flavor).

Change-Id: I749f73b1705f8e4f9af3d02788a0effd0270a6e4
Signed-off-by: Michal Berger <michal.berger@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/19560
Reviewed-by: Jim Harris <jim.harris@gmail.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>