vhost-scsi: add new rpc to enhance live recovery
After the vhost-scsi controller is started, qemu can reconnect
to spdk, but at this time the LUN may not be added to the scsi
target, struct spdk_scsi_dev_vhost_state's member dev has not
been set yet, which will lead to IO error subsequently.

So we add a vhost_create_scsi_controller RPC option to create
controller but doesn't start it, and add a new RPC to start
the controller separately.

Change-Id: I1f8a78528de3073293d1c5ce2c064123519bc320
Signed-off-by: Haoqian He <haoqian.he@smartx.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/19590
Reviewed-by: Changpeng Liu <changpeng.liu@intel.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Jim Harris <jim.harris@samsung.com>
Community-CI: Mellanox Build Bot