iobuf: always dump config
Previously, the config was only dumped, if the iobuf_set_options RPC has
been received.  It used to work that way, because when iobuf was
initially introduced, these options could also be configured via
spdk_bdev_set_opts, for backwards compatibility.  However, this behavior
was removed in v23.05, so the config can be dumped indiscriminately now.

Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: Id2fb4a47ff6da2eb770b1339e9afe6fc18c4f64d
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/19681
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Jim Harris <jim.harris@gmail.com>
Reviewed-by: Krzysztof Karas <krzysztof.karas@intel.com>