test/scheduler: Don't declare $plugin as ro
Bash does not differentiate read only variables in global scope
from the local instances. Currently conflict exists with the
fio_plugin() which does not allow scheduler/common.sh to be
source'ed into env where this function is meant to be used.

To mitigate, just not declare scheduler/common.sh's $plugin
as read-only.

Change-Id: Ia1a569fc41bd90df580cd7267bd1f29480286278
Signed-off-by: Michal Berger <michal.berger@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/19757
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Jim Harris <jim.harris@gmail.com>
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>