lib/ftl: Abandonment of FTL VSS EMU
Vss emulation was meant to allow for executing FTL on top of non-vss
drives with some caveats (not being power failure safe being the main
issue). However the following patches are adding the logic necessary to
run a fully functional FTL on such drives, so the code here is
abandoned.

Change-Id: I7da2af756662aa8549e979a4803848ed913e17e3
Signed-off-by: Mariusz Barczak <Mariusz.Barczak@solidigmtechnology.com>
Signed-off-by: Mateusz Kozlowski <mateusz.kozlowski@solidigm.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/18547
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Ben Walker <ben@nvidia.com>
Reviewed-by: Jim Harris <jim.harris@samsung.com>
Reviewed-by: Artur Paszkiewicz <artur.paszkiewicz@intel.com>