lib/ftl: Add IO error handling in band write
Previous implementation would call the ftl_p2l_ckpt_issue regardless if
IO failed or not, at which point some of the context for retrying may be
lost. Adding the check early, so that the p2l checkpoint can continue
working on the assumption that it's working after a successful band
write.

Change-Id: Ia7683f067f625f28ce8a02b1bf1896629f14344b
Signed-off-by: Mariusz Barczak <Mariusz.Barczak@solidigmtechnology.com>
Signed-off-by: Mateusz Kozlowski <mateusz.kozlowski@solidigm.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/19255
Reviewed-by: Artur Paszkiewicz <artur.paszkiewicz@intel.com>
Reviewed-by: Jim Harris <jim.harris@samsung.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Ben Walker <ben@nvidia.com>