lib/ftl: Handle trim requests without VSS
The trim flow and its metadata are upgraded. To assure trim atomicity
a new metadata layout region is added - trim log. The flow of the trim
is as follows:
1. Log the trim requests
2. Persist the trim metadata
3. Clear the trim log

If power loss occurs during the trim, the trim log is used to restore
the entire trim request.

Change-Id: I132c6bb2b04fbe68a94a38c495344ffd4b8f6d47
Signed-off-by: Mariusz Barczak <Mariusz.Barczak@solidigmtechnology.com>
Signed-off-by: Lukasz Lasek <lukasz.lasek@solidigmtechnology.com>
Signed-off-by: Mateusz Kozlowski <mateusz.kozlowski@solidigm.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/19609
Community-CI: Mellanox Build Bot
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Jim Harris <jim.harris@samsung.com>
Reviewed-by: Ben Walker <ben@nvidia.com>
Reviewed-by: Artur Paszkiewicz <artur.paszkiewicz@intel.com>