lib/ftl: API for superblock layout upgrade
Moved the content of ftl_sb.c to ftl_sb_v3.c file.
The ftl_sb.c is now a wrapper to dispatch the SB version and call the
appropriate implementation.
Upon the change of FTL SB layout, a new implementation will be added for
the new version.

This is necessary as FTL SB upgrades were meant to be incremental.
Changing the layout requires the FTL to be able to reference the SB,
regardless of its version, at runtime.

Change-Id: I2ac955b267ff4abb302b95f2da3678ff81345c18
Signed-off-by: Lukasz Lasek <lukasz.lasek@solidigm.com>
Signed-off-by: Mariusz Barczak <Mariusz.Barczak@solidigmtechnology.com>
Signed-off-by: Mateusz Kozlowski <mateusz.kozlowski@solidigm.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/19594
Reviewed-by: Artur Paszkiewicz <artur.paszkiewicz@intel.com>
Reviewed-by: Jim Harris <jim.harris@samsung.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Ben Walker <ben@nvidia.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>