env_dpdk: fixed potential dereferencing NULL ptr
Fixed issue detected by Coverity.
If 'params' are not NULL, spdk_jsonrpc_send_error_response
is called, which may free 'request' in case of error during
sending (e.g. connection was closed). Then there shall be
return from the rpc_env_dpdk_get_mem_stats() function,
but this was not the case and rest of the code was executed.
Which in turn might cause using of already freed memory.

Removed error log, which was not related to this function.

Change-Id: I5d910fb6c9f2d14b22baa21d051d4d5db20bdd9c
Signed-off-by: Marcin Spiewak <marcin.spiewak@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/19891
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Community-CI: Mellanox Build Bot