bdev/nvme: add option to control accel sequence support
There are still some limitiations to the way IOs with an accel sequence
are handled: mainly, they'll never be retried by bdev_nvme, because it
cannot re-execute an accel sequence (and it doesn't even know whether
that sequence was executed or not).

We plan to address this issue in the future, but, in the interim, a new
option, allow_accel_sequence, was added to allow users to control
whether NVMe bdevs report support for accel sequences (of course the
underlying controller must also support them).

Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: I560419c605b1f6e2f825a33411e06d120c4d5629
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/20014
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Aleksey Marchuk <alexeymar@nvidia.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>