autopackage: Move porcelain check to autobuild
autopackage.sh has no view into autobuild nor autotest infra, hence
it should not be responsible for similar verification. It was also
depending on a wonky config.mk check which was assuming SPDK had
been built prior running this script.

Remove all that and move the check to autobuild as an extra test in
build_files().

If we need to verify if autotest itself is leaving any garbage
behind, then similar functionality must be implemented directly
in the test suite.

Change-Id: If338b16779feeb2d1a7bf7bea51cba8da18ffef7
Signed-off-by: Michal Berger <michal.berger@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/20025
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Karol Latecki <karol.latecki@intel.com>
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>