nvmf: add nvmf_passthru_admin_cmd_for_ctrlr
This is a helper function to enable sending admin
cmds that target a controller rather than a
namespace. In this case the cmd's nsid will not
be valid for looking up a bdev, but we need to send
it through a bdev since there is no controller
entity in the bdev layer. It will just send it
through the first namespace on the controller.

This function will be used in upcoming patch, so
for now do not mark it static to avoid unused
warnings, we will add it when it gets used in next
patch.

Signed-off-by: Jim Harris <jim.harris@samsung.com>
Change-Id: I1d723a928f9e32bc7361028fe7820a52b21151d7
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/20033
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Shuhei Matsumoto <smatsumoto@nvidia.com>
Reviewed-by: Karl Bonde Torp <k.torp@samsung.com>
Reviewed-by: Aleksey Marchuk <alexeymar@nvidia.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Changpeng Liu <changpeng.liu@intel.com>