nvmf: add IDENTIFY_IOCS (cns 1Ch) emulation
Linux kernel sends IDENTIFY CNS 0x1C to check if
ZNS (or KV) command sets are supported, so add this
emulation in the nvmf target code.

Currently it just checks namespaces that are currently
attached to the subsystem, and doesn't account for case
where namespaces are attached later.  But existing
ZNS code in nvmf layer already takes this same approach
for other fields. In the future we may want to consider
RPCs for the subsystem that explicitly enable certain
command sets even if there are not yet any namespaces
attached that support those command sets.

Signed-off-by: Jim Harris <jim.harris@samsung.com>
Change-Id: Ia91f89d8df0b40b4ccf95d11418a58551a32c26b
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/20046
Community-CI: Mellanox Build Bot
Reviewed-by: Karl Bonde Torp <k.torp@samsung.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Reviewed-by: Changpeng Liu <changpeng.liu@intel.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>