vmd: rename vmd_map_bars -> vmd_domain_map_bars
There already is a function called vmd_dev_map_bar(), but it's used in a
different context (it maps BARs of the devices behind the VMD), so
vmd_map_bars() was renamed to better reflect its purpose.

Additionally, this patch removes the spdk_pci_device argument from this
function, as it's available from struct vmd_adapter.

Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: I8de03c159653a04286e71e9adf226eee40ef06ae
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/20186
Community-CI: Mellanox Build Bot
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Jim Harris <jim.harris@samsung.com>
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>