accel: add module priority
We already had a notion of some kind of module priorities, i.e. the
software module had the lowest priority and was always chosen as a
fallback to the hardware modules.  It relied on keeping the software
module first on the list of modules.

This patch makes this mechanism more generic and allows for specifying
the priority of any module.  The software module has a priority of
SPDK_ACCEL_SW_PRIORITY (-1), all hardware modules should set their
priorities to a non-negative value.

It will make it possible to define modules that would only be used for
testing.  Their priority would be set to a value lower than -1, so
they'd never be used unless explicitly assigned to an operation via
spdk_accel_assign_opc().

Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: Ifd5648fe973d9ba68835786ccceba5c8ad6700ec
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/20434
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Aleksey Marchuk <alexeymar@nvidia.com>
Reviewed-by: Jim Harris <jim.harris@samsung.com>
Community-CI: Mellanox Build Bot