accel: add error module
This module will allow users to inject errors to selected accel
operations, similarly to how bdev_error allows user to inject errors on
the bdev layer.

Since the errors won't necessarily be injected to each submitted
operation, the module needs to be able to actually execute some of them
correctly.  To avoid duplicating the code, it forwards the execution of
the tasks to the software module.

This patch only defines the basic module structures, but doesn't inject
any errors.  That will be done in the following patches.

Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: I708e2429cf7ce5373beaf991ccd0be8fc45b9fc0
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/20436
Reviewed-by: Aleksey Marchuk <alexeymar@nvidia.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Jim Harris <jim.harris@samsung.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>