test/nvme/cuse: Return proper nvme-cli's exit status
We don't use pipefail in our autotest infra so the piped sed here
masks potential failures coming from the nvme-cli. To avoid that,
check nvme-cli's exit status explicitly.

This is cosmetic, but it allows for the tests to fail early on
rather than at the end when the final checks are being made.

Change-Id: I94e975b112da1b216666621b035681ae05374aaf
Signed-off-by: Michal Berger <michal.berger@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/20254
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Reviewed-by: Karol Latecki <karol.latecki@intel.com>
Reviewed-by: Jim Harris <jim.harris@samsung.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>