env_dpdk: drop compatibility with DPDK < v21.11
The oldest, still supported, version of DPDK is v20.11 and it'll be
EOL-ed in December, before the next SPDK release, so let's just drop
compatibility with anything older than v21.11 now.  It was half-broken
anyway, as our copies of DPDK headers had an

#ifdef RTE_VERSION < RTE_VERSION_NUM(21, 11, 0, 0)

in one of the structure definitions, so the code built against v20.11
wouldn't work correctly under v21.11+.

Fixes #3153.

Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: I3d3f45382f7858e60ca5f013d0c0e2fd5cd9959f
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/20331
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Reviewed-by: Michal Berger <michal.berger@intel.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Jim Harris <jim.harris@samsung.com>
Community-CI: Mellanox Build Bot