bdev/nvme: bdev_nvme_get_controller_health_info add critical_warning field
In spdk, we can use the rpc command bdev_nvme_get_controller_health_info to obtain disk hardware related information,
but the 'critical_warning' field is missed.

The 'critical_warning' field indicates critical warnings for the state of the controller. This field is very important in production,
which is used to determine whether the disk hardware is healthy.

Change-Id: I4d935ff65a6f7ea1adaf643c2ac523f2b7afbd15
Signed-off-by: luchangqi.123 <luchangqi.123@bytedance.com>
Signed-off-by: zhenwei pi <pizhenwei@bytedance.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/20355
Reviewed-by: Aleksey Marchuk <alexeymar@nvidia.com>
Community-CI: Mellanox Build Bot
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Changpeng Liu <changpeng.liu@intel.com>
Reviewed-by: Jim Harris <jim.harris@samsung.com>